Conflicts:
	config/generateErrorfile.py

+++ arangod/Ahuacatl/ahuacatl-tokens.cpp
@@ -33,7 +33,7 @@
 #if defined (__STDC_VERSION__) && __STDC_VERSION__ >= 199901L
 
 /* C99 says to define __STDC_LIMIT_MACROS before including stdint.h,
- * if you want the limit (max/min) macros for int types.
+ * if you want the limit (max/min) macros for int types. 
  */
 #ifndef __STDC_LIMIT_MACROS
 #define __STDC_LIMIT_MACROS 1
@@ -50,7 +50,7 @@ typedef uint32_t flex_uint32_t;
 typedef signed char flex_int8_t;
 typedef short int flex_int16_t;
 typedef int flex_int32_t;
-typedef unsigned char flex_uint8_t;
+typedef unsigned char flex_uint8_t; 
 typedef unsigned short int flex_uint16_t;
 typedef unsigned int flex_uint32_t;
 
@@ -184,7 +184,7 @@ typedef struct yy_buffer_state *YY_BUFFER_STATE;
 
     /* Note: We specifically omit the test for yy_rule_can_match_eol because it requires
      *       access to the local variable yy_act. Since yyless() is a macro, it would break
-     *       existing scanners that call yyless() from OUTSIDE Ahuacatllex.
+     *       existing scanners that call yyless() from OUTSIDE Ahuacatllex. 
      *       One obvious solution it to make yy_act a global. I tried that, and saw
      *       a 5% performance hit in a non-yylineno scanner, because yy_act is
      *       normally declared as a variable-- so it is not worth it.
@@ -196,7 +196,7 @@ typedef struct yy_buffer_state *YY_BUFFER_STATE;
                     if ( yytext[yyl] == '\n' )\
                         --yylineno;\
             }while(0)
-
+    
 /* Return all but the first "n" matched characters back to the input stream. */
 #define yyless(n) \
 	do \
@@ -258,7 +258,7 @@ struct yy_buffer_state
 
     int yy_bs_lineno; /**< The line count. */
     int yy_bs_column; /**< The column count. */
-
+    
 	/* Whether to try to fill the input buffer when we reach the
 	 * end of it.
 	 */
@@ -608,9 +608,9 @@ static yyconst flex_int16_t yy_chk[452] =
 /* Table of booleans, true if rule could match eol. */
 static yyconst flex_int32_t yy_rule_can_match_eol[79] =
     {   0,
-0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0,
+0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 
+    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 
+    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 
     0, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0,     };
 
 /* The intent behind this definition is that it'll catch
@@ -623,7 +623,7 @@ static yyconst flex_int32_t yy_rule_can_match_eol[79] =
 
 #include "Basics/Common.h"
 
-#include "BasicsC/tri-strings.h"
+#include "Basics/tri-strings.h"
 
 #include "Ahuacatl/ahuacatl-ast-node.h"
 #include "Ahuacatl/ahuacatl-context.h"
@@ -715,9 +715,9 @@ static int yy_init_globals (yyscan_t yyscanner );
     /* This must go here because YYSTYPE and YYLTYPE are included
      * from bison output in section 1.*/
     #    define yylval yyg->yylval_r
-
+    
     #    define yylloc yyg->yylloc_r
-
+    
 int Ahuacatllex_init (yyscan_t* scanner);
 
 int Ahuacatllex_init_extra (YY_EXTRA_TYPE user_defined,yyscan_t* scanner);
@@ -756,9 +756,9 @@ YYSTYPE * Ahuacatlget_lval (yyscan_t yyscanner );
 void Ahuacatlset_lval (YYSTYPE * yylval_param ,yyscan_t yyscanner );
 
        YYLTYPE *Ahuacatlget_lloc (yyscan_t yyscanner );
-
+    
         void Ahuacatlset_lloc (YYLTYPE * yylloc_param ,yyscan_t yyscanner );
-
+    
 /* Macros after this point can all be overridden by user definitions in
  * section 1.
  */
@@ -979,7 +979,7 @@ yy_find_action:
 			int yyl;
 			for ( yyl = 0; yyl < yyleng; ++yyl )
 				if ( yytext[yyl] == '\n' )
-
+					   
     do{ yylineno++;
         yycolumn=0;
     }while(0)
@@ -1035,7 +1035,7 @@ YY_RULE_SETUP
 	YY_BREAK
 case 7:
 YY_RULE_SETUP
-{
+{ 
   return T_LIMIT;
 }
 	YY_BREAK
@@ -1234,7 +1234,7 @@ YY_RULE_SETUP
 case 39:
 YY_RULE_SETUP
 {
-  return T_RANGE;
+  return T_RANGE; 
 }
 	YY_BREAK
 /* ---------------------------------------------------------------------------
@@ -1287,10 +1287,10 @@ YY_RULE_SETUP
   * --------------------------------------------------------------------------- */
 case 47:
 YY_RULE_SETUP
-{
+{ 
   /* unquoted string */
-  yylval->strval = TRI_RegisterStringAql(yyextra, yytext, yyleng, false);
-  return T_STRING;
+  yylval->strval = TRI_RegisterStringAql(yyextra, yytext, yyleng, false); 
+  return T_STRING; 
 }
 	YY_BREAK
 case 48:
@@ -1410,17 +1410,17 @@ YY_RULE_SETUP
 	YY_BREAK
 case 63:
 YY_RULE_SETUP
-{
+{  
   /* a numeric integer value */
-  yylval->strval = TRI_RegisterStringAql(yyextra, yytext, yyleng, false);
+  yylval->strval = TRI_RegisterStringAql(yyextra, yytext, yyleng, false); 
   return T_INTEGER;
 }
 	YY_BREAK
 case 64:
 YY_RULE_SETUP
-{
+{  
   /* a numeric double value */
-  yylval->strval = TRI_RegisterStringAql(yyextra, yytext, yyleng, false);
+  yylval->strval = TRI_RegisterStringAql(yyextra, yytext, yyleng, false); 
   return T_DOUBLE;
 }
 	YY_BREAK
@@ -1432,7 +1432,7 @@ YY_RULE_SETUP
 {
   /* bind parameters must start with a @
      if followed by another @, this is a collection name parameter */
-  yylval->strval = TRI_RegisterStringAql(yyextra, yytext + 1, yyleng - 1, false);
+  yylval->strval = TRI_RegisterStringAql(yyextra, yytext + 1, yyleng - 1, false); 
   return T_PARAMETER;
 }
 	YY_BREAK
@@ -1443,7 +1443,7 @@ case 66:
 /* rule 66 can match eol */
 YY_RULE_SETUP
 {
-  /* whitespace is ignored */
+  /* whitespace is ignored */ 
 }
 	YY_BREAK
 case 67:
@@ -1469,7 +1469,7 @@ YY_RULE_SETUP
 	YY_BREAK
 case 70:
 YY_RULE_SETUP
-{
+{ 
   // eat comment in chunks
 }
 	YY_BREAK
@@ -1487,7 +1487,7 @@ YY_RULE_SETUP
 	YY_BREAK
 case 73:
 YY_RULE_SETUP
-{
+{ 
   // eat comment in chunks
 }
 	YY_BREAK
@@ -1930,7 +1930,7 @@ static int yy_get_next_buffer (yyscan_t yyscanner)
 	yyg->yy_hold_char = *++yyg->yy_c_buf_p;
 
 	if ( c == '\n' )
-
+		   
     do{ yylineno++;
         yycolumn=0;
     }while(0)
@@ -2013,7 +2013,7 @@ static void Ahuacatl_load_buffer_state  (yyscan_t yyscanner)
     YY_BUFFER_STATE Ahuacatl_create_buffer  (FILE * file, int  size , yyscan_t yyscanner)
 {
 	YY_BUFFER_STATE b;
-
+    
 	b = (YY_BUFFER_STATE) Ahuacatlalloc(sizeof( struct yy_buffer_state ) ,yyscanner );
 	if ( ! b )
 		YY_FATAL_ERROR( "out of dynamic memory in Ahuacatl_create_buffer()" );
@@ -2057,7 +2057,7 @@ static void Ahuacatl_load_buffer_state  (yyscan_t yyscanner)
 #ifndef __cplusplus
 extern int isatty (int );
 #endif /* __cplusplus */
-
+    
 /* Initializes or reinitializes a buffer.
  * This function is sometimes called more than once on the same buffer,
  * such as during a Ahuacatlrestart() or at EOF.
@@ -2083,7 +2083,7 @@ extern int isatty (int );
     }
 
         b->yy_is_interactive = file ? (isatty( fileno(file) ) > 0) : 0;
-
+    
 	errno = oerrno;
 }
 
@@ -2189,9 +2189,9 @@ static void Ahuacatlensure_buffer_stack (yyscan_t yyscanner)
 								, yyscanner);
 		if ( ! yyg->yy_buffer_stack )
 			YY_FATAL_ERROR( "out of dynamic memory in Ahuacatlensure_buffer_stack()" );
-
+								  
 		memset(yyg->yy_buffer_stack, 0, num_to_alloc * sizeof(struct yy_buffer_state*));
-
+				
 		yyg->yy_buffer_stack_max = num_to_alloc;
 		yyg->yy_buffer_stack_top = 0;
 		return;
@@ -2220,12 +2220,12 @@ static void Ahuacatlensure_buffer_stack (yyscan_t yyscanner)
  * @param base the character buffer
  * @param size the size in bytes of the character buffer
  * @param yyscanner The scanner object.
- * @return the newly allocated buffer state object.
+ * @return the newly allocated buffer state object. 
  */
 YY_BUFFER_STATE Ahuacatl_scan_buffer  (char * base, yy_size_t  size , yyscan_t yyscanner)
 {
 	YY_BUFFER_STATE b;
-
+    
 	if ( size < 2 ||
 	     base[size-2] != YY_END_OF_BUFFER_CHAR ||
 	     base[size-1] != YY_END_OF_BUFFER_CHAR )
@@ -2261,7 +2261,7 @@ YY_BUFFER_STATE Ahuacatl_scan_buffer  (char * base, yy_size_t  size , yyscan_t y
  */
 YY_BUFFER_STATE Ahuacatl_scan_string (yyconst char * yystr , yyscan_t yyscanner)
 {
-
+    
 	return Ahuacatl_scan_bytes(yystr,strlen(yystr) ,yyscanner);
 }
 
@@ -2278,14 +2278,14 @@ YY_BUFFER_STATE Ahuacatl_scan_bytes  (yyconst char * yybytes, int  _yybytes_len 
 	char *buf;
 	yy_size_t n;
 	int i;
-
+    
 	/* Get memory for full buffer, including space for trailing EOB's. */
 	n = _yybytes_len + 2;
 	buf = (char *) Ahuacatlalloc(n ,yyscanner );
 	if ( ! buf )
 		YY_FATAL_ERROR( "out of dynamic memory in Ahuacatl_scan_bytes()" );
 
-	for ( i = 0; i < _yybytes_len; ++i )
+	for ( i = 0; i < (int) _yybytes_len; ++i )
 		buf[i] = yybytes[i];
 
 	buf[_yybytes_len] = buf[_yybytes_len+1] = YY_END_OF_BUFFER_CHAR;
@@ -2346,10 +2346,10 @@ YY_EXTRA_TYPE Ahuacatlget_extra  (yyscan_t yyscanner)
 int Ahuacatlget_lineno  (yyscan_t yyscanner)
 {
     struct yyguts_t * yyg = (struct yyguts_t*)yyscanner;
-
+    
         if (! YY_CURRENT_BUFFER)
             return 0;
-
+    
     return yylineno;
 }
 
@@ -2359,10 +2359,10 @@ int Ahuacatlget_lineno  (yyscan_t yyscanner)
 int Ahuacatlget_column  (yyscan_t yyscanner)
 {
     struct yyguts_t * yyg = (struct yyguts_t*)yyscanner;
-
+    
         if (! YY_CURRENT_BUFFER)
             return 0;
-
+    
     return yycolumn;
 }
 
@@ -2423,8 +2423,8 @@ void Ahuacatlset_lineno (int  line_number , yyscan_t yyscanner)
 
         /* lineno is only valid if an input buffer exists. */
         if (! YY_CURRENT_BUFFER )
-           yy_fatal_error( "Ahuacatlset_lineno called with no buffer" , yyscanner);
-
+           yy_fatal_error( "Ahuacatlset_lineno called with no buffer" , yyscanner); 
+    
     yylineno = line_number;
 }
 
@@ -2438,8 +2438,8 @@ void Ahuacatlset_column (int  column_no , yyscan_t yyscanner)
 
         /* column is only valid if an input buffer exists. */
         if (! YY_CURRENT_BUFFER )
-           yy_fatal_error( "Ahuacatlset_column called with no buffer" , yyscanner);
-
+           yy_fatal_error( "Ahuacatlset_column called with no buffer" , yyscanner); 
+    
     yycolumn = column_no;
 }
 
@@ -2492,13 +2492,13 @@ YYLTYPE *Ahuacatlget_lloc  (yyscan_t yyscanner)
     struct yyguts_t * yyg = (struct yyguts_t*)yyscanner;
     return yylloc;
 }
-
+    
 void Ahuacatlset_lloc (YYLTYPE *  yylloc_param , yyscan_t yyscanner)
 {
     struct yyguts_t * yyg = (struct yyguts_t*)yyscanner;
     yylloc = yylloc_param;
 }
-
+    
 /* User-visible API */
 
 /* Ahuacatllex_init is special because it creates the scanner itself, so it is
@@ -2546,20 +2546,20 @@ int Ahuacatllex_init_extra(YY_EXTRA_TYPE yy_user_defined,yyscan_t* ptr_yy_global
         errno = EINVAL;
         return 1;
     }
-
+	
     *ptr_yy_globals = (yyscan_t) Ahuacatlalloc ( sizeof( struct yyguts_t ), &dummy_yyguts );
-
+	
     if (*ptr_yy_globals == NULL){
         errno = ENOMEM;
         return 1;
     }
-
+    
     /* By setting to 0xAA, we expose bugs in
     yy_init_globals. Leave at 0x00 for releases. */
     memset(*ptr_yy_globals,0x00,sizeof(struct yyguts_t));
-
+    
     Ahuacatlset_extra (yy_user_defined, *ptr_yy_globals);
-
+    
     return yy_init_globals ( *ptr_yy_globals );
 }
 
@@ -2674,11 +2674,3 @@ void Ahuacatlfree (void * ptr , yyscan_t yyscanner)
 
 #define YYTABLES_NAME "yytables"
 
-// -----------------------------------------------------------------------------
-// --SECTION--                                                       END-OF-FILE
-// -----------------------------------------------------------------------------
-
-// Local Variables:
-// mode: outline-minor
-// outline-regexp: "/// @brief\\|/// {@inheritDoc}\\|/// @page\\|// --SECTION--\\|/// @\\}"
-// End:
